+++ src/api/words/get-word-by-id.api.ts
@@ -1,6 +1,6 @@
 import axios from 'axios'
 import { CustomizedAxiosResponse } from '../index.interface'
-import { WordData } from './words.interface'
+import { WordData } from './interfaces'
 
 export const getWordByIdApi = async (
   wordId: string,
