+++ src/api/words/get-words.api.ts
@@ -1,6 +1,6 @@
 import axios from 'axios'
 import { CustomizedAxiosResponse } from '../index.interface'
-import { WordData } from './words.interface'
+import { WordData } from './interfaces'
 
 export const getWordsApi = async (): Promise<
   CustomizedAxiosResponse<WordData[]>
