+++ src/api/words/interfaces/index.search-params.ts
@@ -1,0 +1,12 @@
+import { GlobalLanguageCode } from '@/global.interface'
+
+export interface WordSearchParams {
+  id: string
+  languageCode: GlobalLanguageCode
+  semester: number
+  isFavorite: boolean
+  term: string
+  pronunciation: string
+  definition: string
+  example: string
+}
