+++ src/api/words/words.guard.ts
@@ -1,4 +1,4 @@
-import { WordData } from './words.interface'
+import { WordData } from './interfaces'
 
 export const isWordData = (object: unknown): object is WordData => {
   if (!object || typeof object !== `object`) return false
@@ -6,3 +6,5 @@ export const isWordData = (object: unknown): object is WordData => {
   // TODO: More type guard operator?
   return true
 }
+
+// TODO: What is this? Remove this or use it.
