+++ src/dto/index.root.ts
@@ -1,0 +1,3 @@
+export interface GetReqDtoRoot {
+  limit: number
+}
