+++ src/lambdas/parse-input-into-word.lambda.ts
@@ -1,6 +1,6 @@
 // ! Fully tested and confirmed on Jan 8, 2023
 
-import { WordData } from '@/api/words/words.interface'
+import { WordData } from '@/api/words/interfaces'
 import { getRandomHexHandler } from '@/handlers/get-random-hex.handler'
 import { stringSliceHandler } from '@/handlers/string-slice.handler'
